requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.12.14
cffi==1.17.1
colorama==0.4.6
coverage==7.6.12
greenlet==3.1.1
iniconfig==2.0.0
mysql-connector-python==9.2.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-sqlalchemy==0.2.1
python-dotenv==1.0.1
sentry-sdk==2.20.0
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
urllib3==2.3.0
